refactor: Use CSS Modules to style Header.
